refactor: split handler file
