+++ internal/api/handler/hanlder.go
@@ -1,0 +1,13 @@
+package handler
+
+// hanlder.go - ハンドラ系の共通部分
+// (エンドポイント定義無し)
+
+import (
+	domrepo "github.com/ramsesyok/oss-catalog/internal/domain/repository"
+)
+
+type Handler struct {
+	AuditRepo       domrepo.AuditLogRepository
+	ScopePolicyRepo domrepo.ScopePolicyRepository
+}
